Add latest scraper boilerplate
This was taken from the croatia-parliament [1] scraper and stripped back
to just the necessary bits.

It would be nice to create a scraper generator to take care of this
step. It would also be nice to create an Everypolitician::Scraper
library that can encapsulate these last remaining bits of code that are
needed in a new scraper.

[1]: https://github.com/everypolitician-scrapers/croatia-parliament