Add MemberPage class
This is a basic class dealing with a page for a single member. I've
started with this class because it doesn't have any dependencies, so can
be tested in isolation.

Note that this has currently _only_ been tested with the two people in
the test fixtures.

This was largely taken from the previous attempt to port the scraper to
Scraped, but I've removed two fields that didn't make sense in this new
scraper:

- iddisputado - this wasn't actually being used in EveryPolitician
- phone/fax - neither of these were picking anything up